tests/requirements.txt
pytest>=7.0.0
pytest-django>=4.5.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0
httpx>=0.28.0






